Major refactor, added decode/encode pattern
This is a major refactor in preparation for supporting parsing channel
and telemetry contents.  It refactors the interface for decoding types
from the object initialization function (i.e. `__init__`) to a class
function named `decode`.  A method named `encode` to perform the inverse
operation is also added.

"Native types" have been renamed to "fundamental types".

A casualty of the refactor is that fundamental types are no longer just
int and float.  Instead they are explicitly represented by a custom
class with decode/encode.  This means that accessing the actual value
requires an additional step by accessing the `value` member field.

Similarly binary blobs are no longer represented by just a bytes object.
Instead they are now represented by a `Buffer` object that follows the
decode/encode interface.

A namespace of types is introduced in the form of `fptypes`.  Currently
it only contains fundamental types but as types are parsed and created
from the dictionary they will be added to the namespace for global type
lookup.